bugfix: max_samples should be a float, not int

+++ annif/backend/mllm.py
@@ -224,7 +224,7 @@ class MLLMModel:
                     DecisionTreeClassifier(
                         min_samples_leaf=int(params['min_samples_leaf']),
                         max_leaf_nodes=int(params['max_leaf_nodes'])
-                    ), max_samples=int(params['max_samples'])))])
+                    ), max_samples=float(params['max_samples'])))])
         # fit the model on the training corpus
         self._model.fit(train_X, train_y)
